Initial commit!
.Net Core 2.1 AWS Lambda API template with S3 stuff removed, because it's unneeded.

Added a custom dummy controller to check whether the application works - the further development will continue from there.

There's this web.config from a template. Can't tell if it's something important or not for the lambda. I'll leave it as I won't use it to store any sensitive information anyway - there're environment variables for that.