Removed unneccessary mock setups.
Being bit rusty on C#, I jumped the gun a little and added the mock setups for the usecase and gateway to return null result.

Thing is, I didn't need to. As it's only necessary, when the return result of the mock would be used somewhere other than just returning it or passing it around (like control flow).

Realized this by thinking about the validation control flow on the controller, and writing the test for it.