Added ErrorMessagesFormatter helper class
Added it because it makes dealing with error message strings easier by allowing to control them all from one place rather than go about every test and validator file whenever it changes, or whenever I need to add another error message, where only field name differs.

Another advantage of it is that all error messages remain consistent, as we're working with template for specific type of error case rather than individual error message for every case and field.

Made it static mainly because I want to make a call to it from Error response constructor, and I don't want the response object containing this logic. There shouldn't be any downside performance wise as the method is quite lightweight and it's being called only when an error has occured, so we're at the end of an exceptional flow anyway.