Added Controller Bad Request 400 response when the request is invalid.
Changed the validator interface so that I wouldn't need to create multiple copy paste interface file for each validator to be created. This reduces the amount of duplicate code, hence following the DRY principle.

Also added TestHelper method to set up validator mock with returning Success Validation response. Main reason being, I don't want that set up to make it harder to read the test, when it is of no importance to the test. In this case, every place where this helper method is used only has to abide by that setup so that program wouldn't crash with object null ref exception, when validation flow tries to access null validation result. In essence, this set up is not so much related to what's being tested, but rather it's related to the technicalities of trying to keep code running when changing the real validator with an empty shell mock.

Now that the branching based of validation result is present, I added the "usecase does not get called, when request is in invalid" test, as it only makes sense in the context of branching ('usecase always called' vs 'usecase called upon success').