Added Custom Exception Error Handing to the ArrearsAction Controller.
Created an ErrorThrowerHelper to help with randomizing Exceptions and Error Messages for the test set up. Essentially to make the tests more reabable.

Used an Enum to control what kind of errors would be thrown so that the tests could be made more granural. Essentially to TDD out the Catch vs Catch when - they obviousy come with different constraints, hence would have different tests.