Minor Response object changes
Changed the attribute names to the 1st letter uppercase to conform to C# convention.

Changed the name of Errors tests folder into Responses, as I feel the need to test the success response.

Added a test for the GetAractionsByPropRefResponse success response object.

What's left is to later figure out how to change how the names of these properties look in Json, once I get to serialization stuff.