Improve on Entity generation
Used the faker's generation with ruleset feature. It does the job, but it makes my IDE so laggy I have to wait for it to do syntax highlighting and autocomplete for couple seconds. While this looks cool and classy - I'm not going to use this anymore as it's slowing me down.